+++ mod-record-specifications-dto/src/main/java/org/folio/rspec/domain/dto/DefinitionType.java
@@ -1,0 +1,18 @@
+package org.folio.rspec.domain.dto;
+
+import lombok.Getter;
+
+@Getter
+public enum DefinitionType {
+
+  FIELD("field"),
+  INDICATOR("indicator"),
+  INDICATOR_CODE("indicator-code"),
+  SUBFIELD("subfield");
+
+  private final String type;
+
+  DefinitionType(String type) {
+    this.type = type;
+  }
+}