+++ mod-record-specifications-validator/src/main/java/org/folio/rspec/validation/validator/SpecificationValidator.java
@@ -1,0 +1,11 @@
+package org.folio.rspec.validation.validator;
+
+import java.util.List;
+import org.folio.rspec.domain.dto.SpecificationDto;
+import org.folio.rspec.domain.dto.ValidationError;
+
+public interface SpecificationValidator {
+
+  List<ValidationError> validate(Object recordValue, SpecificationDto specification);
+
+}